% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_loader.R
\name{addLoader}
\alias{addLoader}
\title{Add CSS loaders from server}
\value{
CSS load in R6 class
}
\description{
Add/remove CSS loaders from server to any Shiny/HTML component.
It is useful to indicate busy status when some code is running in the server
and when it finishes, remove the loader to indicate clear status.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{addLoader$new()}}
\item \href{#method-show}{\code{addLoader$show()}}
\item \href{#method-hide}{\code{addLoader$hide()}}
\item \href{#method-clone}{\code{addLoader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{addLoader$new(
  target_selector = "",
  isID = TRUE,
  type = "default",
  id = "",
  height = NULL,
  width = height,
  color = "#337ab7",
  opacity = 1,
  method = "replace",
  block = TRUE,
  center = TRUE,
  bg_color = "#eee",
  footer = NULL,
  z_index = 2000,
  alert = FALSE,
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_selector}}{string, which Shiny component you want to add the
loader to? a shiny component ID or a valid CSS selector if \code{isID = FLASE}.
for example, you have a button and want to add animation to it:\preformatted{actionButton(inputId = "btn")
}

This function is used in server only, so if you are in shiny module,
\strong{DO NOT} add the \code{ns()} wrapper.}

\item{\code{isID}}{bool, is your selector an ID?}

\item{\code{type}}{string, one of  "circle", "dual-ring", "facebook", "heart",
"ring", "roller", "default", "ellipsis", "grid", "hourglass", "ripple",
"spinner", default is "default".}

\item{\code{id}}{string, the unqiue ID for the loader, if not provided, a random
ID will be given. If you are using shiny modules, use \code{session$ns('YOUR_ID')}
to wrap it.}

\item{\code{height}}{string, (r)em, "1.5rem", "1.5em", or pixel, like "10px".
Default is \code{NULL}, will be automatically calculated based on the target
component. It is recommend to use \code{NULL} for "replace" and "inline" method
to let it automatically be calculated, but required for "full_screen" method.}

\item{\code{width}}{string, default is the same as \code{height} to make it sequare.}

\item{\code{color}}{string, any valid CSS color name, or hex color code}

\item{\code{opacity}}{number, between 0-1}

\item{\code{method}}{one of "replace", "inline", "full_screen", see details}

\item{\code{block}}{bool, for some input components, once the loader starts,
it can also block user interaction with the component, very useful for
"inline" method, eg. prevent users from clicking the button while some
process is still running.}

\item{\code{center}}{bool, try to place the load to the center of the target for
"inline" and "replace" and center of the screen for "full_screen".}

\item{\code{bg_color}}{string, any valid CSS color name, or hex color code. Only
works for "full_screen" method.}

\item{\code{footer}}{Additional Shiny/HTML component to add below the loader, like
a title \code{h1("load title")}.}

\item{\code{z_index}}{number, only works for "full_screen" method, what CSS layer
should the overlay be places. In HTML, all elements have the default of 0.}

\item{\code{alert}}{bool, should alert if target cannot be found or other javascript
errors? mainly for debugging}

\item{\code{session}}{shiny session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{addLoader$show(alert = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hide"></a>}}
\if{latex}{\out{\hypertarget{method-hide}{}}}
\subsection{Method \code{hide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{addLoader$hide(alert = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{addLoader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
