% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popover_tip.R
\name{bsHoverPopover}
\alias{bsHoverPopover}
\alias{bsPopover}
\title{Enhanced Bootstrap popover}
\usage{
bsHoverPopover(tag, title, content = "", placement = "top", trigger = "hover")

bsPopover(tag, title, content = "", placement = "top", trigger = "hover")
}
\arguments{
\item{tag}{\link[htmltools:tag]{htmltools::tag}, like htmltools::tags$button() or htmltools::tags$a(),
or \code{\link[shiny:actionButton]{shiny::actionButton()}}}

\item{title}{character, title for the popover, generally text}

\item{content}{character, content for the popover body, can be HTML}

\item{placement}{character, placement of the popover with respect to \code{tag}, one
of "top", "right", "bottom", and "left"}

\item{trigger}{trigger method, default "hover", one of click , hover ,
focus , manual.}
}
\value{
shiny element, tag
}
\description{
enhanced Bootstrap 3 popover by hovering, clicking
or other methods. Default is by "hover". Original bootstrap3 method only works with
"click" on buttons.
}
\details{
\link{bsHoverPopover} is the old name of  \link{bsPopover}, the same function.

Read more about popover here: https://getbootstrap.com/docs/3.3/javascript/#popovers
}
\examples{
if(interactive()){
    library(shiny)
    library(magrittr)
    ui <- fluidPage(
        column(2),
        column(
            8,
            actionButton('a', 'On button') \%>\%
                bsPopoverEnhance(
                    title = "title a",
                    content = "popover works on a button",
                    placement = "bottom"
                ),
            tags$a("On link") \%>\%
                bsPopoverEnhance(
                    title = "title b",
                    content = "popover works on a link",
                    placement = "bottom"
                ),
            div(
              tags$b("general element"),
              style =
                '
              height: 100px;
              background-color: cornflowerblue;
            '
            ) \%>\%
              bsPopoverEnhance(
                title = "general element",
                content = "popover works on a 'div'",
                placement = "right"
              )
        )

    )
    server <- function(input, output, session) {}
    shinyApp(ui, server)
}
}
