% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popover_tip.R
\name{bsTip}
\alias{bsTip}
\title{Enhanced Bootstrap tooltip}
\usage{
bsTip(
  tag,
  title = "",
  placement = "top",
  bgcolor = "black",
  textcolor = "white",
  fontsize = "12px",
  trigger = "hover focus"
)
}
\arguments{
\item{tag}{a shiny tag as input}

\item{title}{string, tooltip text}

\item{placement}{string, one of "top", "bottom", "left", "right", where to put the
tooltip}

\item{bgcolor}{string, background color, valid value of CSS color name or hex value or rgb value}

\item{textcolor}{string, text color, valid value of CSS color name or hex value or rgb value}

\item{fontsize}{string, text font size, valid value of CSS font size, like "10px",
"1rem".}

\item{trigger}{string, how to trigger the tooltip, one or combination of
click | hover | focus | manual.}
}
\value{
shiny tag
}
\description{
Add tooltip to any Shiny element you want. You can also customize
color, font size, background color, trigger event for each individual tooltip.
}
\details{
For trigger methods read: https://getbootstrap.com/docs/3.3/javascript/#tooltips-options.
}
\examples{
if(interactive()){
  library(shiny)
  library(magrittr)
  ui <- fluidPage(
    br(), br(), br(), br(), br(), br(), column(2),
    actionButton("a", "Tooltip on the left") \%>\%
      bsTip("Tooltip on the left", "left"),
    actionButton("b", "Tooltip on the top") \%>\%
      bsTip("Tooltip on the top", "top"),
    actionButton("c", "Tooltip on the right") \%>\%
      bsTip("Tooltip on the right", "right"),
    actionButton("d", "Tooltip on the bottom") \%>\%
      bsTip("Tooltip on the bottom", "bottom"),
    br(), br(), column(2),
    actionButton("e", "primary color") \%>\%
      bsTip("primary color", bgcolor = "#0275d8"),
    actionButton("f", "danger color") \%>\%
      bsTip("danger color", bgcolor = "#d9534f"),
    actionButton("g", "warning color") \%>\%
      bsTip("warning color", bgcolor = "#f0ad4e"),
    br(), br(), column(2),
    actionButton("h", "9px") \%>\%
      bsTip("9px", fontsize = "9px"),
    actionButton("i", "14px") \%>\%
      bsTip("14px", fontsize = "14px"),
    actionButton("j", "20px") \%>\%
      bsTip("20px", fontsize = "20px"),
    br(), br(), column(2),
    actionButton("k", "combined") \%>\%
      bsTip(
        "custom tooltip", "bottom",
        "#0275d8", "#eee", "15px"
      )
  )
  server <- function(input, output, session) {}
  shinyApp(ui, server)
}
}
