% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{renderDesc}
\alias{renderDesc}
\title{Render some collapsible markdown text}
\usage{
renderDesc(id, desc)
}
\arguments{
\item{id}{element ID}

\item{desc}{one character string in markdown format}
}
\value{
HTML elements
}
\description{
write some text in markdown format and it will help you
render to markdown, use \link[shiny:markdown]{shiny::markdown} but it is collapsible.
}
\examples{
if(interactive()){
    desc <-
        "
    # Some desc
    - xxxx
    - bbbb

    This is a [link](https://www.google.com/).

    `Some other things`
    > other markdown things

    1. aaa
    2. bbb
    3. ccc
    "
    ui <- fluidPage(
        renderDesc(id = "desc", desc),
    )

    server <- function(input, output, session) {

    }

    shinyApp(ui, server)
}
}
